refactor: Renaming Tab component and move create release note button into tab
